feature/3
Created TestActionException to encapsulate any Exception that occurs due to invalid method or class declarations which have an annotation associated with it. (For example, method annotated with @ActionMethod must take zero parameters and return TestResult)

TestStepFactory makes use of the ClassScanner class to return instances of TestStep based on a String representing the action

TestStepMethodFactory makes use of the MethodScanner class to return Method instances of methods tagged with @ActionMethod based on a String representing the action. This factory returns objects of type TestStepMethod. This is a new (experimental) class that wraps a method to be called as part of a test. This is purely experimental and is not guaranteed to stay